requirements.txt
chromadb
sentence-transformers
google-generativeai